Genome comparision method
Phylogenetic analysis of DNA sequences or whole genomes by Fourier transform: The DNA sequences or genomes are mapped into 2-dimensional (2D) numerical sequences based on
the nucleotide composition of the DNA sequences. Euclidean distances of DNA sequences of different lengths are computed after even scaling the Fourier power spectra of the DNA sequences to the longest length of the sequences.